+++ Src/Adapters/Nancy.Metrics/NancyMetricsEndpointHandler.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
-using Metrics.Visualization;
+using Metrics.Endpoints;
 
 namespace Nancy.Metrics
 {
