+++ Src/Adapters/Owin.Metrics/Middleware/OwinMetricsEndpointHandler.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
-using System.Linq;
-using Metrics.Visualization;
+using Metrics.Endpoints;
 
 namespace Owin.Metrics.Middleware
 {
