+++ Src/Metrics.Tests/Endpoints/MetricsEndpointHandlerTests.cs
@@ -1,8 +1,8 @@
 ﻿using FluentAssertions;
-using Metrics.Visualization;
+using Metrics.Endpoints;
 using Xunit;
 
-namespace Metrics.Tests.Visualization
+namespace Metrics.Tests.Endpoints
 {
     public class MetricsEndpointHandlerTests
     {
