+++ Src/Metrics/Endpoints/FlotWebApp.cs
@@ -4,7 +4,7 @@ using System.IO;
 using System.IO.Compression;
 using System.Reflection;
 
-namespace Metrics.Visualization
+namespace Metrics.Endpoints
 {
     public static class FlotWebApp
     {
