+++ Src/Metrics/Endpoints/MetricsHttpListener.cs
@@ -1,5 +1,4 @@
-﻿
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.IO;
 using System.IO.Compression;
@@ -10,7 +9,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Metrics.Logging;
 
-namespace Metrics.Visualization
+namespace Metrics.Endpoints
 {
     public sealed class MetricsHttpListener : IDisposable
     {
